+++ src/components/Tables/common/SyTableFilter.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 	import { ref, watch, provide } from 'vue'
 	import type { DataTableHeaders, FilterOption, FilterType } from './types'
+	import { filterItems } from './tableFilterUtils'
 	import SyTextField from '@/components/Customs/SyTextField/SyTextField.vue'
 	import SySelect from '@/components/Customs/SySelect/SySelect.vue'
 	import DatePicker from '@/components/DatePicker/DatePicker.vue'
@@ -87,68 +88,6 @@
 
 		// Emit updated filters
 		emit('update:filters', newFilters)
-	}
-
-	// Filter items based on filters
-	function filterItems<T>(items: T[], filters: FilterOption[]): T[] {
-		if (!filters || filters.length === 0) return items
-
-		return items.filter((item) => {
-			return filters.every((filter) => {
-				if (!filter || !filter.key || !filter.value) return true
-
-				const itemValue = item[filter.key]
-
-				// Handle different filter types
-				switch (filter.type) {
-				case 'text':
-					if (typeof itemValue === 'string' && typeof filter.value === 'string') {
-						return itemValue.toLowerCase().includes(filter.value.toLowerCase())
-					}
-					break
-				case 'number':
-					if (typeof itemValue === 'number' && typeof filter.value === 'number') {
-						return itemValue === filter.value
-					}
-					break
-				case 'select':
-					if (Array.isArray(filter.value)) {
-						return filter.value.includes(itemValue)
-					}
-					else {
-						return itemValue === filter.value
-					}
-				case 'date':
-					if (itemValue instanceof Date && filter.value instanceof Date) {
-						return itemValue.getTime() === filter.value.getTime()
-					}
-					break
-				case 'period':
-					if (typeof filter.value === 'object' && filter.value !== null) {
-						const { from, to } = filter.value as { from?: Date, to?: Date }
-						if (itemValue instanceof Date) {
-							const dateValue = itemValue.getTime()
-							if (from && to) {
-								return dateValue >= from.getTime() && dateValue <= to.getTime()
-							}
-							else if (from) {
-								return dateValue >= from.getTime()
-							}
-							else if (to) {
-								return dateValue <= to.getTime()
-							}
-						}
-					}
-					break
-				default:
-					if (typeof itemValue === 'string' && typeof filter.value === 'string') {
-						return itemValue.toLowerCase().includes(filter.value.toLowerCase())
-					}
-				}
-
-				return false
-			})
-		})
 	}
 
 	// Expose the filtering function via provide/inject pattern
